moved client into test directory
