+++ test/client/api.js
@@ -6,10 +6,10 @@ var EventEmitter = require('events').EventEmitter
 var util = require('util')
 
 var hawk = require('hawk')
-var P = require('../promise')
+var P = require('../../promise')
 var request = require('request')
 
-var tokens = require('../tokens')({ trace: function() {}})
+var tokens = require('../../tokens')({ trace: function() {}})
 
 util.inherits(ClientApi, EventEmitter)
 function ClientApi(origin) {
