+++ test/client/index.js
@@ -2,12 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-var P = require('../promise')
+var P = require('../../promise')
 var ClientApi = require('./api')
-var butil = require('../crypto/butil')
-var pbkdf2 = require('../crypto/pbkdf2')
-var hkdf = require('../crypto/hkdf')
-var tokens = require('../tokens')({ trace: function () {}})
+var butil = require('../../crypto/butil')
+var pbkdf2 = require('../../crypto/pbkdf2')
+var hkdf = require('../../crypto/hkdf')
+var tokens = require('../../tokens')({ trace: function () {}})
 
 function Client(origin) {
   this.uid = null
