+++ test/local/verifier_upgrade_tests.js
@@ -5,7 +5,7 @@
 var test = require('../ptaptest')
 var TestServer = require('../test_server')
 var path = require('path')
-var Client = require('../../client')
+var Client = require('../client')
 var log = { trace: console.log }
 
 process.env.CONFIG_FILES = path.join(__dirname, '../config/verifier_version_0.json')
