+++ test/remote/account_login_tests.js
@@ -6,7 +6,7 @@ var test = require('../ptaptest')
 var TestServer = require('../test_server')
 var path = require('path')
 var crypto = require('crypto')
-var Client = require('../../client')
+var Client = require('../client')
 
 process.env.CONFIG_FILES = path.join(__dirname, '../config/account_tests.json')
 var config = require('../../config').root()
