+++ test/remote/flow_tests.js
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 var test = require('../ptaptest')
-var Client = require('../../client')
+var Client = require('../client')
 var TestServer = require('../test_server')
 var path = require('path')
 var jwcrypto = require('jwcrypto')
