+++ test/remote/misc_tests.js
@@ -4,7 +4,7 @@
 
 var test = require('../ptaptest')
 var TestServer = require('../test_server')
-var Client = require('../../client')
+var Client = require('../client')
 var path = require('path')
 var P = require('../../promise')
 var hawk = require('hawk')
